Fix armv7 build on travis with native armv8 machines
We switched to native arm machines on travis for speed reasons.
The python build detected aarch64 as target architecture even
inside the armv7 containers.
This lead to armv7 wheels that were incorrectly labelled and published
as aarch64. The real aarch64 packages where in turn never published
because the name was already taken.